Version 2.0.0

+++ src/System/Linux/Network.php
@@ -49,8 +49,8 @@ class Network extends AbstractNetwork
         $this->channel = (int) $network[4];
         $this->security = $network[7];
         $this->securityFlags = $network[8].' '.$network[9];
-        $this->dbm = $network[6];
-        $this->quality = to_quality((int) $network[6]);
+        $this->dbm = (float) $network[6];
+        $this->quality = (float) to_quality((int) $network[6]);
         $this->frequency = (int) $network[5];
         $this->connected = ($network[0] == self::POSITIVE_CONNECTION_FLAG);
